pyproject.toml
[project]
name = "memorious"
version = "3.0.1"
description = "A minimalistic, recursive web crawling library for Python."
authors = [
    { name = "Organized Crime and Corruption Reporting Project", email = "data@occrp.org" },
    { name = "Data and Research Center", email = "hi@dataresearchcenter.org" },
]
license = "MIT"
readme = "README.rst"
classifiers = [
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
requires-python = ">=3.11,<4"
dependencies = [
    "banal >= 1.0.6, < 2.0.0",
    "lxml[html-clean] (>=6.0.2,<7.0.0)",
    "normality (>=3.0.2,<4.0.0)",
    "python-dateutil (>=2.9.0.post0,<3.0.0)",
    "alephclient (>=2.6.0,<3.0.0)",
    "followthemoney (>=4.3.2,<5.0.0)",
    "dateparser (>=1.2.1,<2.0.0)",
    "stringcase (>=1.2.0,<2.0.0)",
    "legacy-cgi (>=2.6.3,<3.0.0)",
    "rigour (>=1.4.1,<2.0.0)",
    "werkzeug (>=3.1.3,<4.0.0)",
    "openaleph-servicelayer (>=1.23.4,<2.0.0)",
    "ftmq (>=4.3.3,<5.0.0)",
    "ftm-lakehouse @ git+https://github.com/openaleph/ftm-lakehouse.git",
    "anystore @ git+https://github.com/dataresearchcenter/anystore.git",
    "openaleph-procrastinate (>=5.1.3,<6.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "jq (>=1.6.0,<2.0.0)",
    "furl (>=2.1.0,<3.0.0)",
    "jinja2 (>=3.0.0,<4.0.0)",
]

[project.optional-dependencies]
# level = ["plyvel (>=1.5.1,<2.0.0)"]
sql = ["sqlalchemy (>=2.0.36,<3.0.0)"]
postgres = ["sqlalchemy (>=2.0.36,<3.0.0)", "psycopg2 (>=2.9.10,<3.0.0)"]
redis = ["redis (>=4.0.0,<6.0.0)", "fakeredis (>=2.26.2,<3.0.0)"]
ftp = ["requests[security] (>=2.32.3,<3.0.0)", "PySocks == 1.7.1", "requests-ftp (>=0.3.1,<0.4.0)"]

[project.scripts]
memorious = "memorious.cli:cli"

[project.entry-points."memorious.operations"]
memorious = "memorious.cli:main"
# Fetch operations
fetch = "memorious.operations.fetch:fetch"
fetch_extended = "memorious.operations.fetch:fetch_extended"
session = "memorious.operations.fetch:session"
post = "memorious.operations.fetch:post"
post_json = "memorious.operations.fetch:post_json"
post_form = "memorious.operations.fetch:post_form"
# Parse operations
parse = "memorious.operations.parse:parse"
parse_listing = "memorious.operations.parse:parse_listing"
parse_jq = "memorious.operations.parse:parse_jq"
parse_csv = "memorious.operations.parse:parse_csv"
parse_xml = "memorious.operations.parse:parse_xml"
# Clean operations
clean = "memorious.operations.clean:clean"
clean_html = "memorious.operations.clean:clean_html"
# Extract operations
extract = "memorious.operations.extract:extract"
# Regex operations
regex_groups = "memorious.operations.regex:regex_groups"
# Store operations
store = "memorious.operations.store:store"
directory = "memorious.operations.store:directory"
lakehouse = "memorious.operations.store:lakehouse"
cleanup_archive = "memorious.operations.store:cleanup_archive"
# Initializer operations
seed = "memorious.operations.initializers:seed"
tee = "memorious.operations.initializers:tee"
sequence = "memorious.operations.initializers:sequence"
dates = "memorious.operations.initializers:dates"
enumerate = "memorious.operations.initializers:enumerate"
# Debug operations
inspect = "memorious.operations.debug:inspect"
ipdb = "memorious.operations.debug:ipdb"
# WebDAV operations
dav_index = "memorious.operations.webdav:dav_index"
# FTP operations
ftp_fetch = "memorious.operations.ftp:ftp_fetch"
# DocumentCloud operations
documentcloud_query = "memorious.operations.documentcloud:documentcloud_query"
documentcloud_mark_processed = "memorious.operations.documentcloud:documentcloud_mark_processed"
# Aleph operations
aleph_emit = "memorious.operations.aleph:aleph_emit"
aleph_emit_document = "memorious.operations.aleph:aleph_emit_document"
aleph_folder = "memorious.operations.aleph:aleph_folder"
aleph_emit_entity = "memorious.operations.aleph:aleph_emit_entity"
# FTM operations
balkhash_put = "memorious.operations.ftm:ftm_store"
ftm_store = "memorious.operations.ftm:ftm_store"
ftm_load_aleph = "memorious.operations.ftm:ftm_load_aleph"

[project.urls]
Homepage = "https://docs.investigraph.dev/lib/memorious"
Documentation = "https://docs.investigraph.dev/lib/memorious"
Repository = "https://github.com/dataresearchcenter/memorious"
Issues = "https://github.com/dataresearchcenter/memorious/issues"

[tool.poetry.group.dev.dependencies]
pytest = ">=7.4.3,<9.0.0"
pytest-cov = ">=4.1,<7.0"
pytest-env = "^1.1.1"
pytest-httpbin = "^2.1.0"
black = ">=23.11,<25.0"
isort = "^5.12.0"
mypy = "^1.7.0"
pre-commit = "^4.0.1"
flake8 = ">=6.1,<8.0"
ipdb = "^0.13.13"
bump2version = "^1.0.1"
mkdocs = "^1.6.1"
mkdocstrings-python = "^1.16.10"
mkdocs-autorefs = "^1.4.1"
mkdocs-material = "^9.6.13"
mkdocs-click = "^0.9.0"
pytest-mock = "^3.14.1"
sphinx = "^8.2.3"
sphinx-rtd-theme = "^3.0.2"
recommonmark = "^0.7.1"
mkdocs-typer2 = "^0.1.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest_env]
DEBUG = 1
TESTING = 1
MEMORIOUS_CONFIG_PATH = "./tests/testdata/config"
MEMORIOUS_BASE_PATH = "./tests/testdata/data"
MEMORIOUS_TAGS_URI = "memory://"
PROCRASTINATE_SYNC = 1
PROCRASTINATE_DB_URI = "memory://"
PROCRASTINATE_APP = "memorious.tasks.app"
LAKEHOUSE_URI = "./tests/testdata/data"
